+++ tools/rgb2ycbcr.c
@@ -49,9 +49,9 @@
     TIFFSetField(out, tag, v)
 
 #ifndef howmany
-#define howmany(x, y) (((x) + ((y)-1)) / (y))
+#define howmany(x, y) (((uint32_t)(x) + ((uint32_t)(y) - 1U)) / (uint32_t)(y))
 #endif
-#define roundup(x, y) (howmany(x, y) * ((uint32_t)(y)))
+#define roundup(x, y) ((uint32_t)(howmany(x, y)) * ((uint32_t)(y)))
 
 #define LumaRed ycbcrCoeffs[0]
 #define LumaGreen ycbcrCoeffs[1]
@@ -277,7 +277,7 @@ static int cvtRaster(TIFF *tif, uint32_t *raster, uint32_t width,
     uint32_t rnrows = roundup(nrows, vertSubSampling);
 
     cc = (tsize_t)rnrows * rwidth +
-         2 * ((rnrows * rwidth) / (horizSubSampling * vertSubSampling));
+         2 * ((rnrows * rwidth) / (uint32_t)(horizSubSampling * vertSubSampling));
     buf = (unsigned char *)_TIFFmalloc(cc);
     // FIXME unchecked malloc
     for (y = height; (int32_t)y > 0; y -= nrows)
@@ -286,7 +286,7 @@ static int cvtRaster(TIFF *tif, uint32_t *raster, uint32_t width,
         cvtStrip(buf, raster + (y - 1) * width, nr, width);
         nr = roundup(nr, vertSubSampling);
         acc = (tsize_t)nr * rwidth +
-              2 * ((nr * rwidth) / (horizSubSampling * vertSubSampling));
+              2 * ((nr * rwidth) / (uint32_t)(horizSubSampling * vertSubSampling));
         if (!TIFFWriteEncodedStrip(tif, strip++, buf, acc))
         {
             _TIFFfree(buf);
